move filters to index file
